move task monitor init/start from client.login() to newClient()
reason: each login inits/starts a new task monitor ... this is probably
okay, the channel architecture should allow them to run in parallel.
But that wasn't the intent.

logins can happen due to session end, but also due to an aggressive
caller making lots of requests in parallel with a brand new (not logged-in)
client. Terraform acts this way.

It looks like moving the task monitor to newClient() will avoid that problem
without side effects. time will tell.

an alternative:
- don't init the client's taskmon channel at client init
- after login, establish a lock
- defer unlock
- determine whether the taskmon channel is nil
  - no: do nothing
  - yes: create the channel(s); start the taskmon