Eliminate checking the `type` field when unmarshaling each template type
Each template used to check the type in its `UnmarshalJSON()` function. This
turned out to be a problem when I got around to to testing 5-stage (pod-based)
templates because the embedded 3-stage (rack-based) templates representing pods
don't have a `type` element.

I was reluctant to let go of the type check, but there's no way to signal to
`UnmarshalJSON()` that it's operating in a nested context.

My first crack at handling this problem involved partially unpacking the nested
templates, adding the missing `type` attribute, re-packing them to `[]byte` and
then invoking `UnmarshalJSON()` more-or-less as usual.

This merge eliminates the check (which is being done further upstream here:
https://github.com/Juniper/apstra-go-sdk/blob/f843bdc632cccffda75b8c45ea9138f9a18bace1/internal/template/common.go#L101C1-L116C1)
and eliminates the pre-unmarshal touch-ups which the check required.